Nest invoice pages and functionality together.
