+++ experiments/billable/src/app/pages/invoice/DetailPage/InvoiceDetailPage.tsx
@@ -1,11 +1,10 @@
 "use server";
 
-import { Suspense } from "react";
-import { Layout } from "../Layout";
+import { Layout } from "../../Layout";
 
 import { InvoiceForm } from "./InvoiceForm";
-import { RouteContext } from "../../../router";
-import { db } from "../../../db";
+import { RouteContext } from "../../../../router";
+import { db } from "../../../../db";
 
 export type InvoiceItem = {
   description: string,
